HSEARCH-2036 Extracting base class for HSQuery implementations
